Splits send-message into agent and signup routes
