+++ lib/middleware/message-outbound-create.js
@@ -3,13 +3,15 @@
 const Promise = require('bluebird');
 const helpers = require('../helpers');
 
-module.exports = function createOutboundMessage(config) {
+module.exports = function createOutboundMessage(config = {}) {
+  const direction = config.messageDirection || 'outbound-api-send';
+  const postToPlatform = config.shouldPostToPlatform || true;
+
   return (req, res) => req.conversation
-    .createLastOutboundMessage(config.messageDirection, req.outboundMessageText,
-      req.outboundTemplate, req)
+    .createLastOutboundMessage(direction, req.outboundMessageText, req.outboundTemplate, req)
     .then(() => {
       let promise = Promise.resolve();
-      if (config.shouldPostToPlatform) {
+      if (postToPlatform) {
         promise = req.conversation.postLastOutboundMessageToPlatform();
       }
       return promise;
