+++ lib/middleware/messages/agent/front.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const logger = require('../../logger');
-const front = require('../../front');
+const logger = require('../../../logger');
+const front = require('../../../front');
 
 module.exports = function sendSupportMessage() {
   return (req, res, next) => {
