+++ lib/middleware/messages/agent/params.js
@@ -1,8 +1,8 @@
 'use strict';
 
 const crypto = require('crypto');
-const logger = require('../../logger');
-const helpers = require('../../helpers');
+const logger = require('../../../logger');
+const helpers = require('../../../helpers');
 
 /**
  * Validates incoming Front request.
