+++ lib/middleware/messages/signup/campaign.js
@@ -1,8 +1,8 @@
 'use strict';
 
-const gambitCampaigns = require('../../gambit-campaigns');
-const UnprocessibleEntityError = require('../../../app/exceptions/UnprocessibleEntityError');
-const helpers = require('../../helpers');
+const gambitCampaigns = require('../../../gambit-campaigns');
+const UnprocessibleEntityError = require('../../../../app/exceptions/UnprocessibleEntityError');
+const helpers = require('../../../helpers');
 
 module.exports = function sendCampaignTemplate() {
   return (req, res, next) => {
