+++ lib/middleware/messages/signup/params.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const helpers = require('../../helpers');
-const UnprocessibleEntityError = require('../../../app/exceptions/UnprocessibleEntityError');
+const helpers = require('../../../helpers');
+const UnprocessibleEntityError = require('../../../../app/exceptions/UnprocessibleEntityError');
 
 module.exports = function campaignParams() {
   return (req, res, next) => {
